New function addAggregatedLiquidity( ) # ***************** #  Function to create aggregated hierarchical reports from contract reports # ***************** #  addAggregateReport(account,cidReports,reportName) #     This function adds a new aggregated report reportName into the list of #     of reports at each node of the accounts tree. Input parameter cidReports #     is a list keyed by CID of report values for each contract. Each report #     is a numeric vector computed by contract cashflow analysis. All report #     vectors have the same length - determined by timeline of the analysis. #     Flow reports have one less element than status/value reports. Function #     addAggregateReport() computes account node report values recursively: #     (1) nonleaf accounts do a vector sum of their child node reports #     (2) leaf nodes owning at least one actusCID, do a vector sum of the #         cidReports for their assigned contracts #     (3) leaf nodes with no assigned actusContracts generate a report vector #         of the correct length ( for this report type) BUT all NULL vallues
The tree can be printed with a vector liquidity values ( at different report times) for each account and parent accounts the aggregated report of child accounts. The same logic could be used for nominal and Net present value hierarchichal report creation  BUT the tree print function will not print selected elements from a list of reports. Also no way to use a single function for liquidity, nominalValue and netPResentValue because no way for the assigned node attribute nme to be passed in as a string.